Save & Exit key function, won't work in text area
This code will allow Save and Exit to be keyed instead of mouseclicked, but it still won't work in a text area.  I've added code to try to get it to work in the text area, but to no avail.  The element in question, input id=sfData_?_?, seems to change with each record, that is, the question marks appear as different numbers in different records, so I don't know how to account for that in the code, and I haven't found any resources on the web yet to explain that.  The code that I've tried comes from http://stackoverflow.com/questions/195951/change-an-elements-class-with-javascript, and there isn't any mention of any other libraries that would need to be included.  I've tried variations of document.getElementByID("sfData").className += " mousetrap";, both inside the function and outside the function, as can be seen in the comments.  I've tried "sfData_*_*", "sfData", and "sfData(this)", but the result is always the same.  Pushing the '+' button in the text field shows up as a +, and the Save and Exit doesn't work.  In addition, Save and Exit doesn't respond to + outside the text field either.